celery_worker/requirements.txt
celery[redis]==5.4.0
Flask==3.1.0
redis==5.2.1
python-dotenv==1.1.0
gunicorn==23.0.0
